+++ flow-typed/lodash.js
@@ -1,3 +1,3 @@
-declare module 'lodash.debounce' {
+declare module 'lodash/debounce' {
   declare export default function debounce(func: Function, wait: number): Function;
 }