+++ packages/interweave-emoji/CHANGELOG.md
@@ -1,4 +1,4 @@
-# 1.0.0
+# 1.0.0 - 9/25/17
 #### 🎉 Release
 * Initial release!